fix: alert bugs (#360)
