+++ src/components.d.ts
@@ -67,10 +67,6 @@ export namespace Components {
          */
         "customClass"?: string;
         /**
-          * Time taken to dismiss the toast in milliseconds
-         */
-        "delay"?: number;
-        /**
           * Whether the alert has a dismiss button
          */
         "dismissible"?: boolean;
@@ -1489,6 +1485,10 @@ export namespace Components {
          */
         "customClass"?: string;
         /**
+          * Time taken to dismiss the toast in milliseconds
+         */
+        "delay"?: number;
+        /**
           * The position of the toast in the parent container.
          */
         "position"?: ToastPosition;
@@ -2514,10 +2514,6 @@ declare namespace LocalJSX {
           * Custom CSS class to apply to the outer div element.
          */
         "customClass"?: string;
-        /**
-          * Time taken to dismiss the toast in milliseconds
-         */
-        "delay"?: number;
         /**
           * Whether the alert has a dismiss button
          */
@@ -4150,6 +4146,10 @@ declare namespace LocalJSX {
           * Additional classes for custom styling.
          */
         "customClass"?: string;
+        /**
+          * Time taken to dismiss the toast in milliseconds
+         */
+        "delay"?: number;
         /**
           * The position of the toast in the parent container.
          */
