+++ src/components/modus-wc-toast/modus-wc-toast.spec.ts
@@ -2,6 +2,12 @@ import { newSpecPage } from '@stencil/core/testing';
 import { ModusWcToast } from './modus-wc-toast';
 
 describe('modus-wc-toast', () => {
+  afterEach(() => {
+    jest.runOnlyPendingTimers();
+    jest.useRealTimers();
+    jest.restoreAllMocks();
+  });
+
   it('should render with default props', async () => {
     const page = await newSpecPage({
       components: [ModusWcToast],
@@ -91,5 +97,61 @@ describe('modus-wc-toast', () => {
       html: '<modus-wc-toast position="bottom-end"><div>Custom content</div></modus-wc-toast>',
     });
     expect(page.root).toMatchSnapshot();
+  });
+
+  it('should set a new timeout on delayChanged and clear timeout on disconnectedCallback', async () => {
+    const page = await newSpecPage({
+      components: [ModusWcToast],
+      html: '<modus-wc-toast delay="500"><div>Custom content</div></modus-wc-toast>',
+    });
+
+    const component = page.rootInstance as ModusWcToast;
+    jest.useFakeTimers();
+    const setTimeoutSpy = jest.spyOn(globalThis, 'setTimeout');
+    const clearTimeoutSpy = jest.spyOn(globalThis, 'clearTimeout');
+
+    component.delayChanged(1000);
+    expect(setTimeoutSpy).toHaveBeenCalledWith(expect.any(Function), 1000);
+
+    component.delayChanged(1100);
+    expect(clearTimeoutSpy).toHaveBeenCalled();
+    expect(setTimeoutSpy).toHaveBeenCalledWith(expect.any(Function), 1100);
+
+    component.disconnectedCallback();
+    expect(clearTimeoutSpy).toHaveBeenCalled();
+  });
+
+  it('should call dismissElement from timeout in delayChanged and componentDidLoad functions', async () => {
+    const page = await newSpecPage({
+      components: [ModusWcToast],
+      html: '<modus-wc-toast delay="500"><div>Custom content</div></modus-wc-toast>',
+    });
+
+    const component = page.rootInstance as ModusWcToast;
+    jest.useFakeTimers();
+    const dismissElementSpy = jest.spyOn(component, 'dismissElement');
+
+    component.delayChanged(500);
+    jest.runAllTimers();
+
+    expect(dismissElementSpy).toHaveBeenCalled();
+
+    component.componentDidLoad();
+    jest.runAllTimers();
+
+    expect(dismissElementSpy).toHaveBeenCalled();
+  });
+
+  it('should clear timeout on disconnectedCallback', async () => {
+    const page = await newSpecPage({
+      components: [ModusWcToast],
+      html: '<modus-wc-toast delay="500"><div>Custom content</div></modus-wc-toast>',
+    });
+
+    const component = page.rootInstance as ModusWcToast;
+    jest.useFakeTimers();
+    const clearTimeoutSpy = jest.spyOn(globalThis, 'clearTimeout');
+    component.disconnectedCallback();
+    expect(clearTimeoutSpy).toHaveBeenCalled();
   });
 });
